service/Services/entities/src/main/java/org/infy/idp/entities/subscription/OrgInfo.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.idp.entities.subscription;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;
/**
 * Entity to store organization information
 * 
 * @author Infosys
 *
 */
public class OrgInfo {

	@SerializedName("orgName")
	@Expose
	private String orgName;

	@SerializedName("orgAdmin")
	@Expose
	private String orgAdmin;

	@SerializedName("domain")
	@Expose
	private String domain;

	public String getOrgName() {
		return orgName;
	}

	public void setOrgName(String orgName) {
		this.orgName = orgName;
	}

	public String getOrgAdmin() {
		return orgAdmin;
	}

	public void setOrgAdmin(String orgAdmin) {
		this.orgAdmin = orgAdmin;
	}

	public String getDomain() {
		return domain;
	}

	public void setDomain(String domain) {
		this.domain = domain;
	}
}
